+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html);
 however, insignificant breaking changes do not guarantee a major version bump, see the reasoning [here](https://github.com/kyb3r/modmail/issues/319). If you're a plugin developer, note the "BREAKING" section.
 
-# v3.7.14-dev2
+# v3.7.14-dev3
 
 ### Added
 
@@ -14,6 +14,7 @@ however, insignificant breaking changes do not guarantee a major version bump, s
 - `?fareply`, anonymously reply with variables.
 - `anonymous_snippets` config variable to toggle if snippets should be anonymous. ([GH #2905](https://github.com/kyb3r/modmail/issues/2905))
 - `disable_updates` config variable to control if the update command should be disabled or not.
+- `silent_alert_on_mention` to alert mods silently. ([GH #2907](https://github.com/kyb3r/modmail/issues/2907))
 - Support for only the "Server Members" intent.
 
 ### Improved